% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_to_hub_schema.R
\name{coerce_to_hub_schema}
\alias{coerce_to_hub_schema}
\alias{coerce_to_character}
\title{Coerce data.frame/tibble column data types to hub schema data types or character.}
\usage{
coerce_to_hub_schema(
  tbl,
  config_tasks,
  skip_date_coercion = FALSE,
  as_arrow_table = FALSE,
  output_type_id_datatype = c("from_config", "auto", "character", "double", "integer",
    "logical", "Date")
)

coerce_to_character(tbl, as_arrow_table = FALSE)
}
\arguments{
\item{tbl}{a model output data.frame/tibble}

\item{config_tasks}{a list version of the content's of a hub's \code{tasks.json}
config file created using function \code{\link[hubUtils:read_config]{hubUtils::read_config()}}.}

\item{skip_date_coercion}{Logical. Whether to skip coercing dates. This can be faster,
especially for larger \code{tbl}s.}

\item{as_arrow_table}{Logical. Whether to return an arrow table. Defaults to \code{FALSE}.}

\item{output_type_id_datatype}{character string. One of \code{"from_config"}, \code{"auto"},
\code{"character"}, \code{"double"}, \code{"integer"}, \code{"logical"}, \code{"Date"}.
Defaults to \code{"from_config"} which uses the setting in the \code{output_type_id_datatype}
property in the \code{tasks.json} config file if available. If the property is
not set in the config, the argument falls back to \code{"auto"} which determines
the  \code{output_type_id} data type automatically from the \code{tasks.json}
config file as the simplest data type required to represent all output
type ID values across all output types in the hub.
When only point estimate output types (where \code{output_type_id}s are \code{NA},) are
being collected by a hub, the \code{output_type_id} column is assigned a \code{character}
data type when auto-determined.
Other data type values can be used to override automatic determination.
Note that attempting to coerce \code{output_type_id} to a data type that is
not valid for the data (e.g. trying to coerce\code{"character"} values to
\code{"double"}) will likely result in an error or potentially unexpected
behaviour so use with care.}
}
\value{
\code{tbl} with column data types coerced to hub schema data types or character.
if \code{as_arrow_table = TRUE}, output is also converted to arrow table.
}
\description{
Coerce data.frame/tibble column data types to hub schema data types or character.
}
\section{Functions}{
\itemize{
\item \code{coerce_to_hub_schema()}: coerce columns to hub schema data types.

\item \code{coerce_to_character()}: coerce all columns to character

}}
