% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_hub.R
\name{collect_hub}
\alias{collect_hub}
\title{Collect Hub model output data}
\usage{
collect_hub(x, silent = FALSE, ...)
}
\arguments{
\item{x}{a \verb{<hub_connection>/<mod_out_connection>} or \verb{<arrow_dplyr_query>} object.}

\item{silent}{Logical. Whether to suppress message generated if conversion to \code{model_out_tbl} fails.}

\item{...}{Further argument passed on to \code{\link[=as_model_out_tbl]{as_model_out_tbl()}}.}
}
\value{
A \code{model_out_tbl}, unless conversion to \code{model_out_tbl} fails in which case a \code{tibble} is returned.
}
\description{
\code{collect_hub} retrieves data from a \verb{<hub_connection>/<mod_out_connection>}
after executing any \verb{<arrow_dplyr_query>} into a local tibble.
The function also attempts to convert the output to a \code{model_out_tbl} class
object before returning.
}
\examples{
hub_path <- system.file("testhubs/simple", package = "hubUtils")
hub_con <- connect_hub(hub_path)
# Collect all data in a hub
hub_con \%>\% collect_hub()
# Filter data before collecting
hub_con \%>\%
  dplyr::filter(is.na(output_type_id)) \%>\%
  collect_hub()
# Pass arguments to as_model_out_tbl()
dplyr::filter(hub_con, is.na(output_type_id)) \%>\%
  collect_hub(remove_empty = TRUE)
}
