% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_target_timeseries.R
\name{connect_target_timeseries}
\alias{connect_target_timeseries}
\title{Open connection to time-series target data}
\usage{
connect_target_timeseries(hub_path = ".")
}
\arguments{
\item{hub_path}{Path to hub directory. Defaults to current working directory.}
}
\value{
An arrow dataset object of subclass <target_timeseries>.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Open the time-series target data file(s)
in a hub as an arrow dataset.
}
\details{
If the target data is split across multiple files in a \code{time-series} directory,
all files must share the same file format, either csv or parquet.
No other types of flies are currently allowed in a \code{time-series} directory.
}
\examples{
# Clone example hub
tmp_hub_path <- withr::local_tempdir()
example_hub <- "https://github.com/hubverse-org/example-complex-forecast-hub.git"
git2r::clone(url = example_hub, local_path = tmp_hub_path)
# Connect to time-series data
ts_con <- connect_target_timeseries(tmp_hub_path)
ts_con
# Collect all time-series data
ts_con |> dplyr::collect()
# Filter for a specific date before collecting
ts_con |>
  dplyr::filter(date == "2020-01-11") |>
  dplyr::collect()
# Filter for a specific location before collecting
ts_con |>
  dplyr::filter(location == "US") |>
  dplyr::collect()
}
