% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_timeseries_schema.R
\name{create_timeseries_schema}
\alias{create_timeseries_schema}
\title{Create time-series target data file schema}
\usage{
create_timeseries_schema(hub_path)
}
\arguments{
\item{hub_path}{Path to hub directory}
}
\value{
an arrow \verb{<schema>} class object
}
\description{
Create time-series target data file schema
}
\examples{
#' # Clone example hub
tmp_hub_path <- withr::local_tempdir()
example_hub <- "https://github.com/hubverse-org/example-complex-forecast-hub.git"
git2r::clone(url = example_hub, local_path = tmp_hub_path)
# Create target time-series schema
create_timeseries_schema(tmp_hub_path)
}
