% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_model_out_val_grid.R
\name{expand_model_out_val_grid}
\alias{expand_model_out_val_grid}
\title{Create expanded grid of valid task ID and output type value combinations}
\usage{
expand_model_out_val_grid(
  config_tasks,
  round_id,
  required_vals_only = FALSE,
  all_character = FALSE,
  as_arrow_table = FALSE,
  bind_model_tasks = TRUE,
  include_sample_ids = FALSE
)
}
\arguments{
\item{config_tasks}{a list version of the content's of a hub's \code{tasks.json}
config file, accessed through the \code{"config_tasks"} attribute of a \verb{<hub_connection>}
object or function \code{\link[hubUtils:read_config]{hubUtils::read_config()}}.}

\item{round_id}{Character string. Round identifier. If the round is set to
\code{round_id_from_variable: true}, IDs are values of the task ID defined in the round's
\code{round_id} property of \code{config_tasks}.
Otherwise should match round's \code{round_id} value in config. Ignored if hub
contains only a single round.}

\item{required_vals_only}{Logical. Whether to return only combinations of
Task ID and related output type ID required values.}

\item{all_character}{Logical. Whether to return all character column.}

\item{as_arrow_table}{Logical. Whether to return an arrow table. Defaults to \code{FALSE}.}

\item{bind_model_tasks}{Logical. Whether to bind expanded grids of
values from multiple modeling tasks into a single tibble/arrow table or
return a list.}

\item{include_sample_ids}{Logical. Whether to include sample identifiers in
the \code{output_type_id} column.}
}
\value{
If \code{bind_model_tasks = TRUE} (default) a tibble or arrow table
containing all possible task ID and related output type ID
value combinations. If \code{bind_model_tasks = FALSE}, a list containing a
tibble or arrow table for each round modeling task.

Columns are coerced to data types according to the hub schema,
unless \code{all_character = TRUE}. If \code{all_character = TRUE}, all columns are returned as
character which can be faster when large expanded grids are expected.
If \code{required_vals_only = TRUE}, values are limited to the combinations of required
values only.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}} This function has been moved to the \code{hubValidations}
package and renamed to \code{expand_model_out_grid()}.
}
\details{
When a round is set to \code{round_id_from_variable: true},
the value of the task ID from which round IDs are derived (i.e. the task ID
specified in \code{round_id} property of \code{config_tasks}) is set to the value of the
\code{round_id} argument in the returned output.

When sample output types are included in the output and \code{include_sample_ids = TRUE},
the \code{output_type_id} column contains example sample indexes which are useful
for identifying the compound task ID structure of multivariate sampling
distributions in particular, i.e. which combinations of task ID values
represent individual samples.
}
