% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-target_data.R
\name{get_target_path}
\alias{get_target_path}
\title{Get the path(s) to the target data file(s) in the hub directory.}
\usage{
get_target_path(hub_path, target_type = c("time-series", "oracle-output"))
}
\arguments{
\item{hub_path}{Either a character string path to a local Modeling Hub directory
or an object of class \verb{<SubTreeFileSystem>} created using functions \code{\link[=s3_bucket]{s3_bucket()}}
or \code{\link[=gs_bucket]{gs_bucket()}} by providing a string S3 or GCS bucket name or path to a
Modeling Hub directory stored in the cloud.
For more details consult the
\href{https://arrow.apache.org/docs/r/articles/fs.html}{Using cloud storage (S3, GCS)}
in the \code{arrow} package.
The hub must be fully configured with valid \code{admin.json} and \code{tasks.json}
files within the \code{hub-config} directory.}

\item{target_type}{Type of target data to retrieve matching files. One of "time-series" or
"oracle-output". Defaults to "time-series".}
}
\value{
a character vector of path(s) to target data file(s) (in the \code{target-data} directory) that make the
\code{target_type} requested.
}
\description{
Get the path(s) to the target data file(s) in the hub directory.
}
\examples{
hub_path <- system.file("testhubs/v5/target_file", package = "hubUtils")
get_target_path(hub_path)
get_target_path(hub_path, "time-series")
get_target_path(hub_path, "oracle-output")
# Access cloud data
s3_bucket_name <- get_s3_bucket_name(hub_path)
s3_hub_path <- s3_bucket(s3_bucket_name)
get_target_path(s3_hub_path)
get_target_path(s3_hub_path, "oracle-output")
}
