% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubData-package.R
\docType{package}
\name{hubData-package}
\alias{hubData}
\alias{hubData-package}
\title{hubData: Tools for accessing and working with hubverse data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of utility functions for accessing and working with forecast and target data from Infectious Disease Modeling Hubs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hubverse-org.github.io/hubData}
  \item \url{https://github.com/hubverse-org/hubData}
  \item Report bugs at \url{https://github.com/hubverse-org/hubData/issues}
}

}
\author{
\strong{Maintainer}: Anna Krystalli \email{annakrystalli@googlemail.com} (\href{https://orcid.org/0000-0002-2378-4915}{ORCID})

Other contributors:
\itemize{
  \item Li Shandross \email{lshandross@umass.edu} [contributor]
  \item Nicholas G. Reich \email{nick@umass.edu} (\href{https://orcid.org/0000-0003-3503-9899}{ORCID}) [contributor]
  \item Evan L. Ray \email{elray@umass.edu} [contributor]
  \item Becky Sweger \email{bsweger@gmail.com} [contributor]
  \item  Consortium of Infectious Disease Modeling Hubs [copyright holder]
}

}
\keyword{internal}
