% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_forecasts_zoltar.R
\name{load_forecasts_zoltar}
\alias{load_forecasts_zoltar}
\title{Load forecasts from zoltardata.com in hubverse format}
\usage{
load_forecasts_zoltar(
  project_name,
  models = NULL,
  timezeros = NULL,
  units = NULL,
  targets = NULL,
  types = NULL,
  as_of = NULL,
  point_output_type = "median"
)
}
\arguments{
\item{project_name}{a string naming the Zoltar project to load forecasts from. assumes the host is zoltardata.com}

\item{models}{a character vector that specifies the models to query. must be model abbreviations. defaults to NULL,
which queries all models in the project}

\item{timezeros}{a character vector that specifies the timezeros to query. must be yyyy-mm-dd format. defaults to
NULL, which queries all timezeros in the project}

\item{units}{a character vector that specifies the units to query. must be unit abbreviations. defaults to NULL,
which queries all units in the project}

\item{targets}{a character vector that specifies the targets to query. must be target names. defaults to NULL,
which queries all targets in the project}

\item{types}{a character vector that specifies the forecast types to query. Choices are "bin", "point", "sample",
"quantile", "mean", and "median". defaults to NULL, which queries all types in the project. note: while Zoltar
supports "named" and "mode" forecasts, this function ignores them}

\item{as_of}{a datetime string that specifies the forecast version. The as_of field format must be a datetime as
parsed by the \href{https://dateutil.readthedocs.io/}{dateutil python library}, which accepts a variety of styles.
You can find examples \href{https://dateutil.readthedocs.io/en/stable/examples.html#parse-examples}{here}.
importantly, the datetime must include timezone information for disambiguation, without which the query will fail.
also, the datatime parsing function used below is extremely lenient when it comes to formatting, so please exercise
caution. defaults to NULL to load the latest version}

\item{point_output_type}{a string that specifies how to convert zoltar \code{point} forecast data to hubverse output type.
must be either "median" or "mean". defaults to "median"}
}
\value{
a hubverse model_out_tbl containing the following columns: "model_id", "timezero", "season", "unit",
"horizon", "target", "output_type", "output_type_id", and "value"
}
\description{
Load forecasts from zoltardata.com in hubverse format
}
\details{
notes:
- requires Z_USERNAME and Z_PASSWORD environment vars
- while Zoltar supports "named" and "mode" forecasts, this function ignores them
- rows with non-numeric values are ignored
- this function removes numeric_horizon mentions from zoltar target names. target names can contain a maximum of
one numeric_horizon. example: "1 wk ahead inc case" -> "wk ahead inc case"
}
\examples{
\dontrun{
df <- load_forecasts_zoltar("Docs Example Project")
df <-
  load_forecasts_zoltar("Docs Example Project", models = c("docs_mod"), timezeros = c("2011-10-16"),
                        units = c("loc1", "loc3"), targets = c("pct next week", "cases next week"),
                        types = c("point"), as_of = NULL, point_output_type = "mean")
}

}
