% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-hubUtils.R
\name{model_id_merge}
\alias{model_id_merge}
\title{Merge model output tbl \code{model_id} column}
\value{
\code{tbl} with either \code{team_abbr} and \code{model_abbr} merged into a single \code{model_id}
column or \code{model_id} split into columns \code{team_abbr} and \code{model_abbr}.

a \link[tibble:tibble]{tibble} with \code{model_id} column split into separate
\code{team_abbr} and \code{model_abbr} columns
}
\description{
See \code{\link[hubUtils:model_id_merge]{hubUtils::model_id_merge()}} for details.
}
\examples{
if (requireNamespace("hubData", quietly = TRUE)) {
  hub_con <- hubData::connect_hub(system.file("testhubs/flusight", package = "hubUtils"))
  tbl <- hub_con \%>\%
    dplyr::filter(output_type == "quantile", location == "US") \%>\%
    dplyr::collect() \%>\%
    dplyr::filter(forecast_date == max(forecast_date))

  tbl_split <- model_id_split(tbl)
  tbl_split

  # Merge model_id
  tbl_merged <- model_id_merge(tbl_split)
  tbl_merged

  # Split / Merge using custom separator
  tbl_sep <- tbl
  tbl_sep$model_id <- gsub("-", "_", tbl_sep$model_id)
  tbl_sep <- model_id_split(tbl_sep, sep = "_")
  tbl_sep
  tbl_sep <- model_id_merge(tbl_sep, sep = "_")
  tbl_sep
}
}
