% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-hubUtils.R
\name{validate_model_out_tbl}
\alias{validate_model_out_tbl}
\title{Validate a \code{model_out_tbl} object}
\value{
If valid, returns a \code{tbl}. Otherwise throws an error.
}
\description{
See \code{\link[hubUtils:validate_model_out_tbl]{hubUtils::validate_model_out_tbl()}} for details.
}
\examples{
library(dplyr)
hub_path <- system.file("testhubs/flusight", package = "hubUtils")
hub_con <- hubData::connect_hub(hub_path)
md_out <- hub_con \%>\%
  filter(output_type == "quantile", location == "US") \%>\%
  collect() \%>\%
  filter(forecast_date == max(forecast_date)) \%>\%
  as_model_out_tbl()

validate_model_out_tbl(md_out)
}
