% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-hubUtils.R
\name{as_model_out_tbl}
\alias{as_model_out_tbl}
\title{Convert model output to a \code{model_out_tbl} class object}
\value{
A \code{model_out_tbl} class object.
}
\description{
See \code{\link[hubUtils:as_model_out_tbl]{hubUtils::as_model_out_tbl()}} for details.
}
\examples{
library(dplyr)
hub_path <- system.file("testhubs/flusight", package = "hubUtils")
hub_con <- hubData::connect_hub(hub_path)
hub_con \%>\%
  filter(output_type == "quantile", location == "US") \%>\%
  collect() \%>\%
  filter(forecast_date == max(forecast_date)) \%>\%
  as_model_out_tbl()
}
