% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cloud.R
\name{get_s3_bucket_name}
\alias{get_s3_bucket_name}
\title{Get the bucket name for the cloud storage location.}
\usage{
get_s3_bucket_name(hub_path = ".")
}
\arguments{
\item{hub_path}{Path to a hub directory.}
}
\value{
The bucket name for the cloud storage location.
}
\description{
Get the bucket name for the cloud storage location.
}
\examples{
hub_path <- hub_path <- system.file("testhubs/v5/target_file", package = "hubUtils")
get_s3_bucket_name(hub_path)
# Get config info from GitHub
get_s3_bucket_name(
  "https://github.com/hubverse-org/example-complex-forecast-hub"
)
}
