% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-hubUtils.R
\name{validate_model_out_tbl}
\alias{validate_model_out_tbl}
\title{Validate a \code{model_out_tbl} object}
\value{
If valid, returns a \code{model_out_tbl} class object. Otherwise, throws an
error.
}
\description{
See \code{\link[hubUtils:validate_model_out_tbl]{hubUtils::validate_model_out_tbl()}} for details.
}
\examples{
md_out <- as_model_out_tbl(hub_con_output)
validate_model_out_tbl(md_out)
}
